refactor(ses): extract SesPolicyService from SesService (strangler step 4)
Fourth store-based domain split, showing the two-way
relationship with the not-yet-extracted Identity domain: the facade keeps
requireIdentityExists and runs it before delegating the v2 policy mutators,
while deleteIdentity now calls policyService.deletePoliciesForIdentity — the
monolith depending on the extracted service. Policy storage, the
per-identity count limit, and name validation move into a new
@ApplicationScoped SesPolicyService. Consumers (v1 SesQueryHandler, v2
SesController) are unchanged.
