+++ src/main/java/io/github/hectorvent/floci/services/ses/SesService.java
@@ -95,9 +95,9 @@ public class SesService {
     private final StorageBackend<String, DedicatedIpPool> dedicatedIpPoolStore;
     private final StorageBackend<String, ContactList> contactListStore;
     private final StorageBackend<String, Contact> contactStore;
-    // Identity (sending authorization) policies: key policy::<region>::<identity>::<policyName>,
-    // value the (normalized) policy JSON. One store shared by the v1 and v2 policy APIs.
-    private final StorageBackend<String, String> policyStore;
+    // Identity (sending authorization) policy storage, extracted to SesPolicyService.
+    // The facade keeps the identity-existence check and delegates the rest.
+    private final SesPolicyService policyService;
     // Receipt-rule-set domain, extracted to its own service. The facade delegates.
     private final SesReceiptRuleService receiptRuleService;
     // Custom verification email templates: storage extracted to SesCvetService. The
@@ -108,9 +108,6 @@ public class SesService {
     // Serializes contact create/update against contact-list deletion so a concurrent delete
     // can't purge the list between validation and the write, leaving an orphaned contact.
     private final Object contactMutationLock = new Object();
-    // Serializes the per-identity policy count check-then-put so concurrent creates can't both pass.
-    private final Object policyMutationLock = new Object();
-    static final int MAX_POLICIES_PER_IDENTITY = 20;
     private final SmtpRelay smtpRelay;
     private final ObjectMapper objectMapper;
     private final SesEventPublisher eventPublisher;
@@ -125,7 +122,7 @@ public class SesService {
     @Inject
     public SesService(StorageFactory storageFactory, SesReceiptRuleService receiptRuleService,
                        SesAccountService accountService, SesCvetService cvetService,
-                       SmtpRelay smtpRelay, ObjectMapper objectMapper,
+                       SesPolicyService policyService, SmtpRelay smtpRelay, ObjectMapper objectMapper,
                        SesEventPublisher eventPublisher, EmulatorConfig config, Route53Service route53Service,
                        Clock clock) {
         this.identityStore = storageFactory.create("ses", "ses-identities.json",
@@ -149,8 +146,7 @@ public class SesService {
                 new TypeReference<Map<String, ContactList>>() {});
         this.contactStore = storageFactory.create("ses", "ses-contacts.json",
                 new TypeReference<Map<String, Contact>>() {});
-        this.policyStore = storageFactory.create("ses", "ses-identity-policies.json",
-                new TypeReference<Map<String, String>>() {});
+        this.policyService = policyService;
         this.receiptRuleService = receiptRuleService;
         this.cvetService = cvetService;
         this.smtpRelay = smtpRelay;
@@ -173,7 +169,7 @@ public class SesService {
                StorageBackend<String, DedicatedIpPool> dedicatedIpPoolStore,
                StorageBackend<String, ContactList> contactListStore,
                StorageBackend<String, Contact> contactStore,
-               StorageBackend<String, String> policyStore,
+               SesPolicyService policyService,
                SesReceiptRuleService receiptRuleService,
                SesCvetService cvetService,
                SmtpRelay smtpRelay,
@@ -181,7 +177,7 @@ public class SesService {
                Clock clock) {
         this(identityStore, emailStore, accountService, templateStore, configSetStore, suppressionStore,
                 accountSuppressionStore, accountVdmStore, dedicatedIpPoolStore, contactListStore, contactStore,
-                policyStore, receiptRuleService, cvetService, smtpRelay, objectMapper, null, clock);
+                policyService, receiptRuleService, cvetService, smtpRelay, objectMapper, null, clock);
     }
 
     SesService(StorageBackend<String, Identity> identityStore,
@@ -195,7 +191,7 @@ public class SesService {
                StorageBackend<String, DedicatedIpPool> dedicatedIpPoolStore,
                StorageBackend<String, ContactList> contactListStore,
                StorageBackend<String, Contact> contactStore,
-               StorageBackend<String, String> policyStore,
+               SesPolicyService policyService,
                SesReceiptRuleService receiptRuleService,
                SesCvetService cvetService,
                SmtpRelay smtpRelay,
@@ -213,7 +209,7 @@ public class SesService {
         this.dedicatedIpPoolStore = dedicatedIpPoolStore;
         this.contactListStore = contactListStore;
         this.contactStore = contactStore;
-        this.policyStore = policyStore;
+        this.policyService = policyService;
         this.receiptRuleService = receiptRuleService;
         this.cvetService = cvetService;
         this.smtpRelay = smtpRelay;
@@ -283,14 +279,7 @@ public class SesService {
 
         // Policies are sub-resources of the identity; drop them too so they can't resurrect into a
         // same-named identity recreated later (and so the per-identity count stays correct).
-        synchronized (policyMutationLock) {
-            String policyPrefix = policyPrefix(region, identityValue);
-            for (String policyKey : new ArrayList<>(policyStore.keys())) {
-                if (policyKey.startsWith(policyPrefix)) {
-                    policyStore.delete(policyKey);
-                }
-            }
-        }
+        policyService.deletePoliciesForIdentity(identityValue, region);
 
         LOG.infov("Deleted identity: {0}", identityValue);
     }
@@ -1906,112 +1895,43 @@ public class SesService {
     // against real AWS. Floci stores and returns policies but does not enforce the authorization
     // (Principal-account existence, Resource-ARN match, or send-time checks) — it has no account
     // registry and does not gate sending, so these are treated as metadata.
-    private static final Pattern POLICY_NAME_CHARS = Pattern.compile("[A-Za-z0-9_-]+");
+    // Policy storage lives in SesPolicyService; this facade forwards, and for the v2 mutators it runs
+    // the identity-existence check (an Identity-domain read) first, before delegating.
 
-    // v1 PutIdentityPolicy: upsert (create or overwrite); v1 does not require the identity to exist.
     public void putIdentityPolicy(String identity, String policyName, String policy, String region) {
-        validatePolicyName(policyName);
-        String normalized = normalizePolicy(policy);
-        String key = policyKey(region, identity, policyName);
-        synchronized (policyMutationLock) {
-            if (policyStore.get(key).isEmpty()) {
-                enforcePolicyLimit(region, identity, false);
-            }
-            policyStore.put(key, normalized);
-        }
-        LOG.infov("SES PutIdentityPolicy: {0} on {1} (region {2})", policyName, identity, region);
+        policyService.putIdentityPolicy(identity, policyName, policy, region);
     }
 
-    // v2 CreateEmailIdentityPolicy: fails if the name already exists; requires the identity.
     public void createEmailIdentityPolicy(String identity, String policyName, String policy, String region) {
         requireIdentityExists(identity, region);
-        validatePolicyName(policyName);
-        String normalized = normalizePolicy(policy);
-        String key = policyKey(region, identity, policyName);
-        synchronized (policyMutationLock) {
-            if (policyStore.get(key).isPresent()) {
-                throw new AwsException("AlreadyExistsException",
-                        "Policy <" + policyName + "> already exists", 400);
-            }
-            enforcePolicyLimit(region, identity, true);
-            policyStore.put(key, normalized);
-        }
-        LOG.infov("SES v2 CreateEmailIdentityPolicy: {0} on {1} (region {2})", policyName, identity, region);
+        policyService.createEmailIdentityPolicy(identity, policyName, policy, region);
     }
 
-    // v2 UpdateEmailIdentityPolicy: fails if the name is missing; requires the identity.
     public void updateEmailIdentityPolicy(String identity, String policyName, String policy, String region) {
         requireIdentityExists(identity, region);
-        validatePolicyName(policyName);
-        String normalized = normalizePolicy(policy);
-        String key = policyKey(region, identity, policyName);
-        synchronized (policyMutationLock) {
-            if (policyStore.get(key).isEmpty()) {
-                throw policyNotFound(policyName);
-            }
-            policyStore.put(key, normalized);
-        }
-        LOG.infov("SES v2 UpdateEmailIdentityPolicy: {0} on {1} (region {2})", policyName, identity, region);
+        policyService.updateEmailIdentityPolicy(identity, policyName, policy, region);
     }
 
-    // v2 GetEmailIdentityPolicies: all policies for the identity; requires the identity.
     public Map<String, String> getEmailIdentityPolicies(String identity, String region) {
         requireIdentityExists(identity, region);
-        return listPolicies(region, identity);
+        return policyService.listAllPolicies(identity, region);
     }
 
-    // v1 GetIdentityPolicies: requested names only, missing silently omitted; no identity check.
     public Map<String, String> getIdentityPolicies(String identity, List<String> policyNames, String region) {
-        Map<String, String> out = new LinkedHashMap<>();
-        for (String name : policyNames) {
-            policyStore.get(policyKey(region, identity, name)).ifPresent(doc -> out.put(name, doc));
-        }
-        return out;
+        return policyService.getIdentityPolicies(identity, policyNames, region);
     }
 
-    // v1 ListIdentityPolicies: policy names (sorted); no identity check.
     public List<String> listIdentityPolicyNames(String identity, String region) {
-        return listPolicies(region, identity).keySet().stream().sorted().toList();
+        return policyService.listIdentityPolicyNames(identity, region);
     }
 
-    // v2 DeleteEmailIdentityPolicy: requires the identity; NotFound if the policy is missing.
     public void deleteEmailIdentityPolicy(String identity, String policyName, String region) {
         requireIdentityExists(identity, region);
-        String key = policyKey(region, identity, policyName);
-        synchronized (policyMutationLock) {
-            if (policyStore.get(key).isEmpty()) {
-                throw policyNotFound(policyName);
-            }
-            policyStore.delete(key);
-        }
-        LOG.infov("SES v2 DeleteEmailIdentityPolicy: {0} on {1} (region {2})", policyName, identity, region);
+        policyService.deleteEmailIdentityPolicy(identity, policyName, region);
     }
 
-    // v1 DeleteIdentityPolicy: idempotent; no identity check, no error on a missing policy.
     public void deleteIdentityPolicy(String identity, String policyName, String region) {
-        policyStore.delete(policyKey(region, identity, policyName));
-        LOG.infov("SES DeleteIdentityPolicy: {0} on {1} (region {2})", policyName, identity, region);
-    }
-
-    private Map<String, String> listPolicies(String region, String identity) {
-        String prefix = policyPrefix(region, identity);
-        Map<String, String> out = new LinkedHashMap<>();
-        for (String key : policyStore.keys()) {
-            if (key.startsWith(prefix)) {
-                policyStore.get(key).ifPresent(doc -> out.put(key.substring(prefix.length()), doc));
-            }
-        }
-        return out;
-    }
-
-    private void enforcePolicyLimit(String region, String identity, boolean v2) {
-        long count = policyStore.keys().stream()
-                .filter(k -> k.startsWith(policyPrefix(region, identity))).count();
-        if (count >= MAX_POLICIES_PER_IDENTITY) {
-            String msg = "Number of policies for <" + identity
-                    + "> exceeds max allowed number of policies per resource";
-            throw new AwsException(v2 ? "LimitExceededException" : "InvalidParameterValue", msg, 400);
-        }
+        policyService.deleteIdentityPolicy(identity, policyName, region);
     }
 
     private void requireIdentityExists(String identity, String region) {
@@ -2019,50 +1939,6 @@ public class SesService {
             throw new AwsException("NotFoundException",
                     "Email identity <" + identity + "> does not exist.", 404);
         }
-    }
-
-    // Error codes are the v1 (Query) codes verified against AWS; the v2 controller remaps them to
-    // BadRequestException. The messages are identical across v1 and v2.
-    private static void validatePolicyName(String policyName) {
-        if (policyName == null || policyName.isBlank()) {
-            throw new AwsException("InvalidParameterValue", "PolicyName is required.", 400);
-        }
-        if (policyName.length() > 64) {
-            throw new AwsException("ValidationError",
-                    "1 validation error detected: Value at 'policyName' failed to satisfy constraint: "
-                            + "Member must have length less than or equal to 64", 400);
-        }
-        if (!POLICY_NAME_CHARS.matcher(policyName).matches()) {
-            throw new AwsException("InvalidParameterValue",
-                    "PolicyName is invalid. Policy names must only include alpha-numeric characters, "
-                            + "dashes, and underscores.", 400);
-        }
-    }
-
-    private String normalizePolicy(String policy) {
-        if (policy == null || policy.isBlank()) {
-            throw new AwsException("InvalidParameterValue", "Policy is required.", 400);
-        }
-        try {
-            // AWS returns the policy with insignificant whitespace stripped; compact it to match.
-            return objectMapper.readTree(policy).toString();
-        } catch (com.fasterxml.jackson.core.JsonProcessingException e) {
-            // Floci does not validate policy semantics; keep an unparseable document verbatim.
-            LOG.debugv("Identity policy is not valid JSON, storing as-is: {0}", e.getMessage());
-            return policy;
-        }
-    }
-
-    private static AwsException policyNotFound(String policyName) {
-        return new AwsException("NotFoundException", "Policy <" + policyName + "> does not exist", 404);
-    }
-
-    private static String policyKey(String region, String identity, String policyName) {
-        return policyPrefix(region, identity) + policyName;
-    }
-
-    private static String policyPrefix(String region, String identity) {
-        return "policy::" + region + "::" + identity + "::";
     }
 
     // Validates Create/Update input in the same two-phase order as real AWS (verified by probe):
