+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceLegacyDkimTokensTest.java
@@ -50,7 +50,7 @@ class SesServiceLegacyDkimTokensTest {
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new InMemoryStorage<String, ContactList>(),
                 new InMemoryStorage<String, Contact>(),
-                new InMemoryStorage<String, String>(),
+                new SesPolicyService(new InMemoryStorage<String, String>(), new ObjectMapper()),
                 new SesReceiptRuleService(new InMemoryStorage<String, ReceiptRuleSet>(), Clock.systemUTC()),
                 new SesCvetService(new InMemoryStorage<String, CustomVerificationEmailTemplate>()),
                 mock(SmtpRelay.class),
