+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceListManagementTest.java
@@ -70,7 +70,7 @@ class SesServiceListManagementTest {
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new InMemoryStorage<String, ContactList>(),
                 contactStore,
-                new InMemoryStorage<String, String>(),
+                new SesPolicyService(new InMemoryStorage<String, String>(), new ObjectMapper()),
                 new SesReceiptRuleService(new InMemoryStorage<String, ReceiptRuleSet>(), Clock.systemUTC()),
                 new SesCvetService(new InMemoryStorage<String, CustomVerificationEmailTemplate>()),
                 smtpRelay,
