+++ src/test/java/io/github/hectorvent/floci/services/ses/SesServiceSmtpTest.java
@@ -49,7 +49,7 @@ class SesServiceSmtpTest {
                 new InMemoryStorage<String, DedicatedIpPool>(),
                 new InMemoryStorage<String, ContactList>(),
                 new InMemoryStorage<String, Contact>(),
-                new InMemoryStorage<String, String>(),
+                new SesPolicyService(new InMemoryStorage<String, String>(), new ObjectMapper()),
                 new SesReceiptRuleService(new InMemoryStorage<String, ReceiptRuleSet>(), Clock.systemUTC()),
                 new SesCvetService(new InMemoryStorage<String, CustomVerificationEmailTemplate>()),
                 smtpRelay,
